requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
httpx
Pillow
python-multipart
